+++ sealed_writer/lib/src/utils/type_utils.dart
@@ -46,17 +46,25 @@ extension TypeUtils on String {
     return true;
   }
 
-  /// can be any time like double or with
-  /// nullability sign like double?.
-  bool isSimpleOrNullableTypeName() {
-    if (isEmpty || trim() != this || contains(' ')) return false;
+  /// can be any thing with/without nullability sign.
+  ///
+  /// can be functional type or alias type name or ...
+  bool isExternalTypeName() {
+    if (isEmpty || trim() != this || '?' == this) return false;
+    return true;
+  }
+
+  /// like isExternalTypeName but disallow nullable types
+  bool isExternalNonNullableTypeName() {
+    if (!isExternalTypeName()) return false;
+    if (endsWith('?')) return false;
     return true;
   }
 
   /// read non-null or nullable type.
   /// for example: double or double?.
   ManifestType readType() {
-    check(isSimpleOrNullableTypeName());
+    check(isExternalTypeName());
     return endsWith('?')
         ? ManifestType(name: substring(0, length - 1), isNullable: true)
         : ManifestType(name: this, isNullable: false);